fix: rename graph folder
