+++ app/src/components/graph/CurationGraph.tsx
