+++ app/src/components/graph/CurationNode.tsx
