+++ app/src/components/graph/WalkthroughNode.tsx
