+++ app/src/components/graph/WalkthroughPanel.tsx
